프로젝트 수정

+++ SPM/src/main/java/project/SPM/service/impl/InfoService.java
@@ -5,10 +5,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import project.SPM.Entity.UserEntity;
 import project.SPM.dto.UserDTO;
+import project.SPM.dto.ViewCarDTO;
+import project.SPM.mapper.ICarMapper;
+import project.SPM.mapper.ICheckMapper;
 import project.SPM.repository.IUserRepository;
 import project.SPM.repository.impl.UserRepository;
 import project.SPM.service.IInfoService;
 import project.SPM.util.EncryptUtil;
+
+import java.util.List;
 
 @Slf4j
 @Service
@@ -16,6 +21,8 @@ import project.SPM.util.EncryptUtil;
 public class InfoService implements IInfoService {
 
     private final IUserRepository iUserRepository;
+    private final ICarMapper iCarMapper;
+    private final ICheckMapper iCheckMapper;
 
     @Override
     public boolean updateInfo(UserDTO userDTO) throws Exception {
@@ -40,8 +47,20 @@ public class InfoService implements IInfoService {
     @Override
     public boolean deleteUser(UserEntity userEntity) throws Exception {
 
-        iUserRepository.delete(userEntity);
-        boolean res = true;
+        boolean res = false;
+
+
+        UserDTO userDTO = new UserDTO();
+        userDTO.setUserId(userEntity.getUserId());
+
+        res = iCarMapper.dropCar(userDTO);
+
+        if (res == true) {
+
+            List<ViewCarDTO> viewCarDTO= iCheckMapper.viewCheck(userDTO);
+            iCheckMapper.dropUser(viewCarDTO);
+            iUserRepository.delete(userEntity);
+        }
 
         return res;
     }
